I did something that would be useful, improved the Contributing.odt file and a bit more things
C++ is now on the limelight!
I shifted functions to C++, and C is now only for a few things. C++ is more efficent for stuff for a browser, and after the CLI is complete I would use Qt for the GUI.

Added links to the functions in the main.c file, and yeah everything is great...

Pranjal Prasad 31 October 2024 18:40