A lot of things are now there.
A lot of things are now there, I have used libxml2 for xml rendering stuff. This is a really big commit for me, this maybe a new starting for the project, libxml2 would have a really good use for the browser engine.
Improved the project tree structure and yeah still I have not written the Manuals entirely.
Commited on 2 November 2024 23:56 GMT +5:30